Refactor of the pgenv_check_dependencies function.
As requested <https://github.com/theory/pgenv/pull/4> the function now prints
all missing dependencies at the end. To achieve this, dependency status
is added on a single line of a temporary file. At the end of the function,
if any missing (i.e., 'KO') line is found it is printed and the script
terminates, otherwise if in verbose mode the whole list of depdendencies
is printed.