Tweak layout, add docs, improve error handling.
Turned out that the program prematurely exited if `which` failed to
find a dependency -- which is not what we want. So switch from `set
-e` to `trap 'exit' ERR` to exit on errors (including function and
subprocess errors, thanks to `set -E`) and disable that auto-exiting
before calling `which` by setting `trap "" ERR`.

Also, document the `check` command and tweak the layout and output
wording.