Fix problem with pgenv_check_dependencies: temporary file not created.
Since the function requires the file to be there in order to perform the check,
the file is created (as temporary) as soon as the script starts. This is
safe since the file will be automatically removed.

Also use a single variable as download path for getting the source
and the list of available releases.