Refactoring of the `available` command to use bash hash.
Since mktemp has diffent behaviors between Linux and Mac OSX, as suggested
in discussion <https://github.com/theory/pgenv/pull/5#issuecomment-415021164>
it is better to rely only on variables and not on temp files.

Therefore this commit removes the PGENV_TEMP_DIR and exploits an hash
indexed by the major version and containing a single row with all versions
separated by a dash.
The major version is multipled by a ten factor, so version 10 becomes 100
and the hash appears to be
    versions_hash[ 100 ] = "10.1-10.2-10.3-10.4-10.5"
    versions_hash[ 96  ] = "9.6.0-9.6.1-9.6.2-9.6.3- ..."

On the output side, the major version is normalized back again and the giant
string is then split into separated values to be printed.

Warning: this commit changes the output order, now most recent versions
are reported as last.