Fix download links for older versions.
Old versions of PostgreSQL were distributed as .tar.gz, not as .tar.bz2.
While it is possible to download always a .tar.gz, let's just assume
that all versions prior to 7 are downloaded as .tar.gz (around 7.4 bz2
tarballs were present too), and newer versions as .tar.bz2.

Also fixes an unclosed parentheses and a problem with the temporary file.