Minor changes to the versions function, does not change command output.
The function pgversions now counts how many versions have been found,
and in case there is nothing installed claims with a message to the user
instead of printing nothing.
Moreover, the output is now done with a single echo and a flags variable.
The output format has not been changed.