Wait for startup.
In pg_ctl the wait option '-w' is the default for stopping, not for
starting. This leads to the program to print a succesful startup message
even if the cluster did not started.
This commits removes the error trapping and waits for pg_ctl
to report startup, then printing a more detailed message about
log inspection.