Fix "unknown command" message.
If the program is invoked without any command, the case enters the default
branch, which shows a warning message using $1. But with no command, $1
has no value at all, so the user sees a message like:

      % pgenv
      Unknown command ``
      ...

Now the program checks if any argument/command has been passed, and if so
it prints an error message, otherwise it does just print the help message.

The help message is printed in any case.