Exploit pgenv_debug when skipping a major version due to user filter.
With the implementation of `available` the user can specify a list of
major versions to show, and this helps cutting the size of the command output.
Since a debug function has been introduced to solve other `available` problems,
it is possible to exploit to print a debug message when throwing away a
major version.

% PGENV_DEBUG=true pgenv available 9.5
...
[DEBUG] Skip major version 9.2 (required 9.5)
[DEBUG] Skip major version 9.3 (required 9.5)
[DEBUG] Skip major version 9.4 (required 9.5)
==============================================
         PostgreSQL major version 9.5
----------------------------------------------
 9.5.0   9.5.1   9.5.2   9.5.3   9.5.4   9.5.5
 9.5.6   9.5.7   9.5.8   9.5.9   9.5.10  9.5.11
 9.5.12  9.5.13  9.5.14
==============================================

[DEBUG] Skip major version 9.6 (required 9.5)
[DEBUG] Skip major version 10 (required 9.5)
[DEBUG] Skip major version 11 (required 9.5)