Do not reference absolute Bash path in she-bang.
In systems where Bash is not installed (or linked) to `/bin/bash`
the she-bang leads to a "bad interpreter" error and makes the script
unusable.
Changing the she-bang to `env` allows the system to find
the correct Bash interpreter on systems where it is installed out of the
usual Linux path (e.g., FreeBSD -> /usr/bin/bash).