Changes to pgversions output (command `versions`).
The version command outputs something like 'pgsql-10.5' but the
required version number when dealing with other commands, like `use`
is '10.5'. To ease the copy-and-paste of version numbers, now `pgversions`
shows on each line the version number and the specific installation:

$ pgenv use 1.08
PostgreSQL 1.08 not installed; installed versions:
        10.5            pgsql-10.5
 *      9.6.10          pgsql-9.6.10

This changes the command output!