When the user has to insert a version number, it can use a beta version.
The regular expression to check for a valid version number did not take into
account the beta versions, that are usually written as
        <major>beta<number>
Since we are currently on 11beta3, and therefore it is possible to assume
that no older a.b major version will be released as beta (!), the regular
expression can consider only 11 or greater beta versions. This means
that something like 9.5beta1 is managed as invalid by default.

Also anchor the end of the regular expression to avoid misleading
version numbers.