Improvements to current selected version management.
There is now a function to check that an instance is currently in use,
and in the case it is not the script is aborted with the usual claim message.
In the case an instance is selected, the version number is printed to STDOUT
and can therefore be captured by those commands that operate against
a particular version (e.g., start, stop, restart).

Messages to the user now reflect the version they are working on.