Refactor of command `version`.
Since it is now possible to get the numeric version for the currently
linked instance, via pgenv_current_postgresql_or_exit, it is possible to
implement the `version` command by means of this internal function.
This also makes a little more consisten the versioning with the
output of the refatored `versions` command.