Refactor `clear` to take advantage of current version function.
Since pgenv_current_postgresql_or_exit checks for a current version
linked, it is possible to use it as a common entry point for those commands
where there is the need to check for the link to exist.
Moreover, since the function exits the script in the case a version is not in
use, there is no other need to check for the link, that can be removed
immediatly reducing the code for implementing the command.