Skeleton for managing PL dependencies.
Now the pgenv_detail_command function accepts an optional
third argument which is managedas a "language":
it allows for checking the existance of the interpreter
to build a PL language at configure time.
Missing languages will not abort the script, but need to change
the way configure will work to avoid failures.
There is also another internal function, defined and invoked by pgenv_check,
named pgenv_configure_pl which accepts a language to use and
build an internal variable to be used as a configure variable
to auto-detect the languages to build PostgreSQL with.