Use a configure variable to make `configure` build PL languages.
After the check for dependencies, a PGENV_CONFIGURE_PL variable is set
with the appropriate variables for building PLPerl and/or PL/Python.
Using such variable, the system will build the instance with the availables
languages (or none).
This also removes the dependency on Perl 5 without breaking compatibility
with prior version.