Remove aggressive check about database administrator.
Reverts commit e6fb0191ca533931499ee102bc16a6d85537e001, that introduced
a function to check for PGENV_USER to exist as system user. Since that
user is used as a database administrator, it is not mandatory to have it
and the aim was to make it simulate the behavior of other OS distributions
that create an user (also to own the daemon).
See comments <https://github.com/theory/pgenv/pull/7#discussion_r214082993>