Do not print installed versions when `build` has no version.
See <https://github.com/theory/pgenv/pull/7#discussion_r214084106>.
When the user wants to build a new instance, and does not specify a
version number at all, it does not make sense to show the list of
already installed PostgreSQL instances (she wants a new one, after all).
The function pgenve_input_version_or_exit has been refactored to accept
a second argument that, when not-empty, produces the list of
already installed versions. Due to this, since the first argument (the user
input) could be empty, it is required to quote every function argument.