Minor changes in the `available` command implementation.
Use an inner conditional to find out if there is the need to print
spaces to aling the first version number in the row. Thsi elimintaes the
duplicated `echo`s.
Also minor changes to the help of the command to make it clear
its output is always sorted without any regard to the user input.