Allow the user to decide which version of PL/Perl and PL/Python to build.
Using the special variables PGENV_PERL and PGENV_PYTHON it is possible to
drive the build process for those embedded languages. The idea is that:
- if no variable is set, the script tries to figure out where the
  interpreter is and automatically builds the langauge if the interpreter
  is found;
- if the variable is set to an executable file, that is used to build the
  language;
- if the variable is set to a non-executable (e.g. 'no', '/dev/null')
  the language will not be built.

This provides a one-shot configuration mechanism to help the user drive the
build process.
See comment <https://github.com/theory/pgenv/pull/9>

The `check` command has been changed so that, in the case a language is
not found, it does suggest the user to set up PGENV_PERL or PGENV_PYTHON
variables as an hint.

Also fixed some typos here and there in the comments.

Added the `env` dependency in the documentation, see ticket #10