Fix internal functions.
The pgenv_current_postgresql_or_exit has been splitted into two functions:
pgenv_current_postgresql_or_exit and pgenv_current_version_number. The former
checks if a version is in use, or it aborts the script with a claiming message.
The latest prints out the current version number.
The split was necessary because the output of the former function was
captured into a variable to be used as a version number to work with,
and so commands like `start`, `stop` and `restart` did fail silently
if no version was in use (the error message was never printed to the user
since it was captured by a variable).

The pgversions function now redirects the `ls` output to sterr, so in the case
there is no instance at all it does not prints an error message, the function
does it by itself.
The function also accepts an optional third argument used as an "hint": if
no version has been built, the command can hint the user about the next
step to do.

Added a few pgenv_debug to place debug messages here and there.