Check for a valid version number in `remove`.
See discussion
<https://github.com/theory/pgenv/commit/3df7c66f7a520d44583f00206e428d9c047f0666>
The idea is to check the user input for a valid input number, but NOT
for an installed version because a not-built (buidl failure) version
could leave tarball and build directory around.

Also remove useless branch in pgenv_input_version_or_exit since
the list of installed versions is handled by another function.