Implementation of the "stop modes" for commands `stop` and `restart`.
It is now possible to specify to the `stop` command a second argument
that is one of the supported modes for pg_ctl:
- smart (the default)
- immediate
- fast
Abbreviations are also recognized (it is possible to specify only the first
letter of the command). If no argument is specified, smart mode is
applied.

In order to allow `pg_ctl` to use its own default stop mode, do not
use 'smart' when the user does not specify a stop mode. To achieve this,
a special variable PGENV_STOP_MODE handles the whole flag for `pg_ctl`
or nothing (default).

The variable PGENV_STOP_MODE can be passed to the script to set the
stop mode, so the following are same invocations:
   $ pgenv stop fast
   $ PGENV_STOP_MODE="fast" pgenv stop

The command line argument takes the precedence on the environment variable.
The environment variable does accept only the 'imemdiate, fast, smart' words
and nothing else, so it cannot be used to inject flags into pg_ctl.

Documentation updated.

Close #12