Make `build` to honor the configuration instead of default runtime.
The `build` command now loads first the configuration, if any, and
then checks for dependencies. In this way, with the new improved dependency
checking, each internal variable (i.e., dependency) is checked only
if not already set in the configuration, otherwise the configuration wins.

Moreover, unified Perl and Python variables so that they use the very same name
in both configuration and depenency checking, and therefore allowing
them to be overwritten.