Handle PL/Perl and PL/Python configuration variables.
When the user writes the configuration explicitly (i.e., `config write`)
the runtime environment is checked with the 'pgenv_check_dependencies()',
so that make, tar, and most notably PL/Perl and PL/Python
variables and configure flags are set up properly.
The 'PGENV_CONFIGURE_PL' variable has been removed to let only
'PGENV_CONFIGURE_OPTS' to include the languages options.

Placed a warning in the configuration comments to let the user know
that 'PGENV_CONFIGURE_OPTS' should not include languages nor prefix
directory.

Removed the 'PGENV_PG_USER' variable: it could clash with the PGENV_INITDB_OPTS
variable, so set it as default and let the user to specify it in the
configuration file.

Minor changes to the quotes.
Updated the documentation.

See comment <https://github.com/theory/pgenv/pull/14#issuecomment-420634414>