Quote the arguments to `configure`.
Apparently the configure command line was built fine, and worked if thrown into
a command prompt, but passing more arguments to `configure` resulted in
an 'invalid package' error. This even if the interpreters were not
specified (i.e., '--with-perl --with-python'
reports 'invalid package: perl --with-python').
Quoting the whole string to `configure` fixes the problem.

Close #15