Tuning of configure option variable.
After a build, the PGENV_CONFIGURE_OPTS could contain PL/Perl and PL/Python
values, that should not be included in the configuration. For this reason,
at the time, the `config write` command zeroes the string so that another
build or an after build will not double the language variables.