Use PGENV_SED in `available`.
Requires to load executable.
Also check for `sort` and `tr` when performing a `check`. Such commands
are used for implementing `available`, but I did not place them into
`PGENV_` variables since I do believe that would lead to a less readable
script and I don't see the point to push to that extent (sed is different,
because it could be simulated by another program, as an instance).