Implementation of a patching feature.
Now there are a couple of pgenv_patch_xxx functions that aims to automate
the patching of the currently building instance. The idea is that patches
are contained in a directory tree rooted at 'patch', and each file corresponds
to a single patch to apply. There are "indexes" that list, in the right order,
which patch files to apply to every version.

Indexes are choosen in a "desperate" manner:
1) exact PostgreSQL version and OS name;
2) exact PostgreSQL version;
3) brand/year and OS name;
4) brand/year;
5) brand and OS name;
6) brand.

So for instance the PostgreSQL version 8.1.4 will apply the first index
found between:

patch/index/patch.8.1.4.Linux
patch/index/patch.8.1.4
patch/index/patch.8.1.Linux
patch/index/patch.8.1
patch/index/patch.8.Linux
patch/index/patch.8

(on a Linux machine).
The first index found readable is the one that wins, so the order the above
list is produced does matter.

The function that performs the patching extracts from the index the name
of a file to apply thru pathc(1). The file name can be absolute or relative,
in such case it must be relative to `patch/` subfolder. A verbose message
is displayed when patching is applied.

So far, if the patching fails, the build stops.