Created a function to display messages at different levels.
There is now a `pgenv_message` functions that accepts a message and a level
(as a text string) and displays the message (on stderr) only if the current
level is the sameor greater of the specified level.

Message levels are kept in a globa array `msg_levels` used as an associative
array where the message level (e.g., 'info') is the key, and the value is a
numeric value. Each time a message is passed to `pgenv_message` the
string value is converted to a numeric value and if this is greater or equal
to the setting PGENV_VERBOSITY the message will be printed.

Advantages: it makes all the messages formatted in the same manner, and
it prints all the messages to stderr (good for redirection).

All calls to removed function `pgenv_debug` has been modified
as `pgenv_message ... 'debug'`.

Also converted the initial call to show PGENV_ROOT to a call of level 'warn'.
See issue #21

Replaced all the "interactive" occurencies of 'echo ' to 'pgenv_message'.