Implementation of 'build-git' command.
This command acts as an alias for 'build dev', that is introduces
the special version 'dev' to instrument the build process to download
the source tree via git instead of getting a tarball.

If the 'dev' version has never been created, than a complete checkout
is performed; otherwise a fetch is done.
The build process continues as usual, with patching, make, and etc.

At the end, the 'pgenv-dev' directory will be in place to indicate the
new version of PostgreSQL.

It is possible to configure the git process via two special variables, that
have not been introduced in the configuration mechanism because they
represent developer tweaks; they can be manually added in any case.

Documentation updated.

See #25