Abort `config show` if no version specified and none in use.
In the case the user has `clear`ed the current instance, so there is no
instance in use, and the `config show` is invoked without any argument
the command cannot show any configuration at all. Therefore, abort the user
about no active PostgreSQL version and exits, otherwise if a version is
in use show its configuration.