First partial implementation of the `alias` command.
The idea is that this command allows the manipulation of PostgreSQL alias
names, stored as links into an `/alias` directory in the `PGENV_ROOT` directory.
Subcommands `add` and `remove` checks for the existance of the links
and create or delete such links int he alias directory.

Changes to the `versions` command in order to show also the aliases.
Now, after the directory output for each version, all aliases are printed out.
This changes the format of the `versions` output.