Implement a `psql` command.
Could be an implementation to avoid problems reported in issue #31.
The problem is that often there is a system-wide installed PostgreSQL
and/or client interfaces (`psql`). When using `psql` the user
gets the system wide that could be incompatible with the installed PostgreSQL
version from pgenv.
Using the `psql` command it is possible to run the version installed along-side
the PostgreSQL version `use`d by pgenv.