Warn the user if the PATH does not include the currently selected PostgreSQL.
See issue #31 and discussion about PR #39.

The idea is that PGENV will check every time a `check`, `use`, `start` command
is issued if the PATH contains $PGENV/pgsql/bin entry and if that
entry is ahead of other entries. In this way, the user will end up using
always the selected PostgreSQL binaries.
A new parameter PGENV_WARNINGS has been introduced in the configuration
settings so that it is possible to avoid the warnings.