Use PGENV_WARNINGS also for warning the user about missing configuration.
When a `build` command is issued, the system prints a message to warn the user
and waits 5 seconds. Since there is a new configuration variable,
PGENV_WARNINGS, it is possible to exclude this message (see #31)
due to configuration.
However, for backward compatibility, the variable is set to on
if no configuration is found.