First version with aliases and multiple installations.
THIS BREAKS BACKWARD COMPATIBILITY.

Now each instance is installed into a directory that includes
a timestamp and is collected into a directory named by the PostgreSQL version,
for instance
 $PGENV_ROOT/
      pgsql-12.0/
              pgsql-12.0-2019-10-21-14-15-35
              pgsql-12.0-2019-10-21-08-36-35
     pgsql-11.5/
              pgsql-11.5-2019-10-20-14-15-35
              pgsql-11.5-2019-10-20-08-36-35

and each instance is automatically assigned two aliases:
- one with the version (12.0 for example)
- one with the full timestamp (pgsql-12.0-2019-10-21-14-15-35).

This means that the alias with the single PostgreSQL version (e.g., 12.0)
works only for the first instance installed.

All other commands have been refactored to use the alias as input.

Remove sed dependency in pgenv_versions.

Fix pgenv_versions to allow '*' on current instance.