Fix to the `clear` command output.
The clear command was not selecting appropriately the number
version of the cluster to stop.
While this has not a side effect, the result was that the configuration file
and the output messages did not catch the concrete PostgreSQL instance.

It was a problem of typing the right function to use.

Fix documentation of the pgenv_stop_or_restart function.