Improved the `log` command.
Now the `log` command displays the currently in use version.
Moreover, it does check that a PostgreSQL version is still in use.
The command did not failed without a version in use, but it is much more
coherent with the rest of the application to check and warn the user.

Exit with the `tail` exit command.

Improved the message in the case there is no log file even if the
database is running.