Susbstitute all the `which` commands to `command -v`.
`command` is a Bash specific way to check and launch a specific
executable. If invoked with -v it provides the full path
to the executable, and it should be faster than invoking `which` in another
process.
Since this script is Bash-dependent, this does not imply any particular
change in the script behavior.