Add PGENV_SCRIPT_PRESTOP.
I realized that usually what you want to execute against a database
is something that involves connecting to the database itself.
Therefore I added the capability to execute a script *before*
the instance is stopped.
The script is protected against failures, so that it does not affect
the pgenv workflow.