Merge branch 'remove_which'.
Susbstitute all the `which` commands to `command -v`.

`command` is a Bash specific way to check and launch a specific
executable. If invoked with -v it provides the full path
to the executable, and it should be faster than invoking `which` in another
process.
Since this script is Bash-dependent, this does not imply any particular
qchange in the script behavior.

Tested against OSX Bash and Ubuntu Bash 5.0.3.

Protect `log` with `command`.

Since `log` executes `tail`, and the user could not have been running
`check` before, it is possible to protect the log action using `command`,
that will fail if no `tail` is in place (that is, in turn, a serious problem
for the user!).

Improve the `config edit` command.

Now there is a check for the EDITOR variable to contain an executable
that can be launched, otherwise the error message is shown.

Close #42