Allow `build` to support `latest` and `earliest` keywords.
It is now possible to specify
build latest
build latest 12
build earliest 12
build earliest

and the system will select automatically the appropriate version
number for download and compilation.

To allow this, I've refactored a piece of code that extracts the hash
of the version numbers into a separated function.

Bump version number.

Improve documentation of internal function.