Verify if pgsql is a link and not manually created
Followed by my previous commit, $PGENV_ROOT/pgsql should be checked if
user has not manually created it as a file or directory.
This could possibly become a source of confusion for user as script
starts to misbehave when $PGENV_ROOT/pgsql/bin or just its parent is
created manually.