Do not warn user when $PGENV_ROOT/pgsql/bin is set in PATH but it does not exist
When "$PGENV_ROOT/pgsql/bin" IS already set in PATH but no instances are in use,
pgenv_warning_path() function still incorrectly warns user that she has not
set that directory in PATH.

This is due to the fact that the condition checks if that entry
points to the same inode that $executables does. So in case the
entry is already set in PATH but it does not exist in $PGENV_ROOT
that condition would return false and misleads user that PATH needs
to be changed.

By the time user executes "pgenv use <version>", that entry will
be created while the entry in PATH is already set, so it is safe to
ignore this condition and return without warning user.