This allows individual items to contain whitespace, which otherwise the shell interpreter splits on, causing issues. For example, this change allows for this configuration:
``` sh
PGENV_CONFIGURE_OPTIONS=(
    --with-perl
    --with-openssl
    'CFLAGS=-I/opt/local/opt/openssl/include -I/opt/local/opt/libxml2/include'
    'LDFLAGS=-L/opt/local/opt/openssl/lib -L/opt/local/opt/libxml2/lib'
)
```

Note the spaces in the `CFLAGS` and `LDFLAGS` options. This did not work when
the options were a string, but do now as an array. The old `_OPTS` names are
still supported, though no longer documented (they get converted to arrays).
Resolves #45.

This PR also contains some cleanup and light editing of the README and removes
commit signature info from the output of `pgenv help`.