Remove verbose message about server not running.
see discussion
<https://github.com/theory/pgenv/pull/53#issuecomment-1120238986>.
If the server is not running and a `stop` command is issued, the
program should not print a verbose message to the terminal indicating
to the user that the server was not runnig. The output should appear
only if the user is in debug mode (PGENV_DEBUG has a value).