Use export instead of -g declare option.
When converting the configuration from the old single value approach
to the new array one, the `typeset` and `declare` builtins are
used. The problem is that the variables must be used as global
variables, so that once the configuration is reloaded, the variable
value is used.
However, seems that `declare - g` is not a valid builtin on Bash on
Mac OSX.
Therefore, when the variable is written to the configuration file, an
explicit `export` for the variable is written too. This ensures that
the variable becomes globally scoped.

Also fix a problem of not-evaluated variable.

Related to #52
Close #54