Enrivonment variable to force the configuration file.
The command now will honor the environment variable
`PGENV_CONFIGURATION_FILE` so that, if set, will determine the
path to the configuration file to be used during all the operations.
This allows to decied which configuration file to use before the
`pgenv` starts working, so for example:

    % export PGENV_CONFIGURATION_FILE=~/git/pgenv/config/default.conf
    % pgenv build 16beta2

will use the default configuration file instead of a version specific
one.

Added also a `pgenv config use <version>` command that prints
the environment variable to export or set in the shell session
(but does not set itself), so that it is possible to do

   % export $( pgenv config use default )

to obtain the environment changed, and hence ease the "computation"
of which file name to use for a specific version.

Close #60