[#62] Command line parsing options
Exploit getopt(1) to get command line options availability.
The idea is to do a first parse at the very beginning of the script,
so that options can be detected and internal and environmental
variables can be set accordingly.
Then the command line parsing continues as before.

Added a new section to the documentation about variables and command
line flags to illustrate how both can be combined and used.

Close #62