Environment variable to prevent config file overwriting.
The variable `PGENV_WRITE_CONFIGURATION_FILE_AUTOMATICALLY`
is used to instrument `pgenv` to decide about the need to
write/overwrite the configuration file at the end
of a build/rebuild process.

If the variable is not set, or set to a true value (i.e., not zero,
not "NO" - case insensitive), the configuration file will be
written and/or overwritten.
If the variable is set to zero or "no" (case insensitive), the
configuration
file will not be written.

Close #62