[#68] Remove old versions from the available list
Since versions 1.x and 6.x have different tarball names and will
probably not compile anymore in an automated way as `pgenv` does,
remove them from the list of available versions for download and
build.

Modifies also the version detecting function, so that in the case a
too old PostgreSQL version is required to be built
(despite the `available` command not showing it),
the program will complain with an error message explaining that
only PostgreSQL versions 7.0 or greater can be managed via
pgenv.

Close #68