[#75] Allows usage of externally managed on-disk PostgreSQL versions
If the user creates a directory in `PGENV_ROOT` with the same layour
of another `pgsql-xxx` directory, and the name with the prefix
`pgsql-`, such directory will be usable by `pgenv use` and subsequent
commands.
This allows, for example, to manage a checked out directory from
another repository via `pgenv` as if it was installed by `pgenv` itself.