Use `world-bin` and install docs separately on 17
PostgreSQL 17 removed pre-compiled docs from the source tarball,
including docs. The docs will now be built from source, but will fail if
the required [DocBook tools] are not installed. To compensate, releases
from 17.0 on are complemented by a docs tarball.

We don't want to add the overhead of requiring the [DocBook tools]. So,
when building Postgres 17 and higher, teach pgenv to use the new
`world-bin` and `install-world-bin` targets to build everything except
non-binary files, which includes docs. Then download and install the
docs tarball and install its contents separately.

Resolves #79.

  [DocBook tools]: https://www.postgresql.org/docs/current/docguide-toolsets.html