[#87] Provide support for major-version based config files
This commit introduces the possibility to handle per-major-version
configuration files (e.g., `17.conf`) that are loaded in the case the
version specific file (.e.g., `17.1.conf`) is not found on the system.

The loading of a configuration file gives priority to the exact
version, so for instance `17.1.conf`, then if such file is not found
the program tries to load a `17.conf` (i.e., major mode file), and at
last the `default.conf` file.

Bump version to 1.3.9

Close #87