Merge configs from least to most specific
Rather than stop at the first configuration found, let each more
specific config override variables from the less specific. The configs
load in this order:

*   `default.conf`
*   `$major.conf`
*   `$major.$minor.conf`

This allows a base level of configuration in `default.conf` and for more
specific configurations to override its values. But keep the first
configuration file loaded in the `PGENV_DEFAULT_CONFIG_FILE` variable so
that it retains the user-specified file if it exists. Resolves #87.