Fix ShellCheck issues, indentation, add workflow
Fix all the issues identified by `shellcheck`, except a few that I'm
unsure about or think should be as they are. Add a GitHub workflow that
runs ShellCheck on `pgenv`. Fix inconsistent indentation and replace
tabs with four spaces. Add `.gitattributes` to exclude files from
archives (such as those generated for tags) and properly increment the
version to v1.4.1.